pyproject.toml
[project]
name = "aws-cost-explorer-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.9",
    "botocore>=1.37.9",
    "jmespath>=1.0.1",
    "mcp[cli]>=1.3.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "tabulate>=0.9.0",
    "requests>=2.31.0",
    "pip>=25.0.1",
    "python-dotenv>=1.0.1",
    "bs4>=0.0.2",
    "readability>=0.3.2",
    "readability-lxml>=0.8.1",
    "lxml>=5.3.1",
    "lxml-html-clean>=0.4.1",
]
